docker/data-collector/requirements.txt
requests
influxdb
schedule
pyyaml
ccxt
pandas
websocket-client
